Apró módosítás a movie Entitásban és a MovieContext-ban

+++ DTOs/RatingsDTO.cs
@@ -4,7 +4,7 @@
     {
        
         public int MovieId { get; set; }
-        public decimal RatingNumber { get; set; }
+        public int RatingNumber { get; set; }
         public string UserId { get; set; }
         public string Comment { get; set; }
     }
